# docs/ARCH/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SPOrTS API",
    "version": "1.0.0",
    "description": "Contratti congelati rispetto all'implementazione attuale (core endpoints)"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/api/global-matches": {
      "get": {
        "summary": "Lista partite globali (admin view)",
        "parameters": [
          {"in":"query","name":"league","schema":{"type":"string"}},
          {"in":"query","name":"fromDate","schema":{"type":"string"}},
          {"in":"query","name":"toDate","schema":{"type":"string"}},
          {"in":"query","name":"limit","schema":{"type":"integer"}},
          {"in":"query","name":"page","schema":{"type":"integer"}}
        ],
        "responses": { "200": {"description": "OK"} }
      }
    },
    "/api/global-matches/leagues": {
      "get": { "summary": "Leghe disponibili", "parameters": [{"in":"query","name":"onlyAvailable","schema":{"type":"boolean"}}], "responses": { "200": {"description": "OK"} } }
    },
    "/api/global-matches/rounds": {
      "get": { "summary": "Partite per giornate (rounds)", "parameters": [{"in":"query","name":"league","required":true,"schema":{"type":"string"}}, {"in":"query","name":"limitRounds","schema":{"type":"integer"}}], "responses": { "200": {"description": "OK"} } }
    },
...
  }
}